#13 refactor: Bid 스크린 모듈화
