+++ RN-mobile/src/components/Bid/BidButton/style.tsx
@@ -1,0 +1,17 @@
+import { StyleSheet } from "react-native";
+
+export const styles = StyleSheet.create({
+  container: {
+    position: "absolute",
+    bottom: 0,
+    right: 0,
+    left: 0,
+    paddingVertical: 18,
+  },
+  bidButton: {
+    fontSize: 20,
+    fontWeight: "500",
+    textAlign: "center",
+    color: "white",
+  },
+});
