+++ RN-mobile/src/components/Bid/PriceInfo/style.tsx
@@ -1,0 +1,18 @@
+import { StyleSheet } from "react-native";
+import { FontColor } from "constants/color";
+
+export const styles = StyleSheet.create({
+  container: {
+    gap: 20,
+  },
+  infoBox: {
+    gap: 10,
+  },
+  subTitle: {
+    fontSize: 16,
+    color: FontColor.gray,
+  },
+  price: {
+    fontWeight: "500",
+  },
+});
